+++ sample/agentidentity/example_agentidentity_create_idp_configuration.go
@@ -1,6 +1,7 @@
 package agentidentitysample
 
 import (
+	"encoding/json"
 	"fmt"
 	"github.com/baidubce/baiducloud-go-sdk/core/util"
 	"github.com/baidubce/baiducloud-go-sdk/services/agentidentity"
@@ -30,9 +31,16 @@ func CreateIdpConfiguration() {
 		DisplayNameClaim:      util.PtrString(""),
 		AutoCreateUser:        util.PtrBool(false),
 	}
-	err = client.CreateIdpConfiguration(createIdpConfigurationRequest)
+	result, err := client.CreateIdpConfiguration(createIdpConfigurationRequest)
 	if err != nil {
 		// 此处仅做打印展示，请谨慎对待异常处理，在工程项目中切勿直接忽略异常。
 		fmt.Println("request failed:", err)
+		return
 	}
+	data, err := json.MarshalIndent(result, "", "    ")
+	if err != nil {
+		fmt.Println("json marshalIndent failed:", err)
+		return
+	}
+	fmt.Println(string(data))
 }