+++ sample/agentidentity/example_agentidentity_create_user_pool.go
@@ -1,6 +1,7 @@
 package agentidentitysample
 
 import (
+	"encoding/json"
 	"fmt"
 	"github.com/baidubce/baiducloud-go-sdk/core/util"
 	"github.com/baidubce/baiducloud-go-sdk/services/agentidentity"
@@ -18,9 +19,16 @@ func CreateUserPool() {
 		Name:        util.PtrString(""),
 		Description: util.PtrString(""),
 	}
-	err = client.CreateUserPool(createUserPoolRequest)
+	result, err := client.CreateUserPool(createUserPoolRequest)
 	if err != nil {
 		// 此处仅做打印展示，请谨慎对待异常处理，在工程项目中切勿直接忽略异常。
 		fmt.Println("request failed:", err)
+		return
 	}
+	data, err := json.MarshalIndent(result, "", "    ")
+	if err != nil {
+		fmt.Println("json marshalIndent failed:", err)
+		return
+	}
+	fmt.Println(string(data))
 }