+++ services/agentidentity/model_get_workload_access_token_response.go
@@ -1,0 +1,9 @@
+package agentidentity
+
+import "github.com/baidubce/baiducloud-go-sdk/bce"
+
+type GetWorkloadAccessTokenResponse struct {
+	bce.BaseResponse
+	Token    *string `json:"token,omitempty"`
+	ExpireAt *string `json:"expireAt,omitempty"`
+}