+++ services/agentidentity/model_list_user_pools_response.go
@@ -4,8 +4,8 @@ import "github.com/baidubce/baiducloud-go-sdk/bce"
 
 type ListUserPoolsResponse struct {
 	bce.BaseResponse
-	TotalCount *int32       `json:"totalCount,omitempty"`
-	PageNo     *int32       `json:"pageNo,omitempty"`
-	PageSize   *int32       `json:"pageSize,omitempty"`
-	Result     *UserPoolDTO `json:"result,omitempty"`
+	TotalCount *int32         `json:"totalCount,omitempty"`
+	PageNo     *int32         `json:"pageNo,omitempty"`
+	PageSize   *int32         `json:"pageSize,omitempty"`
+	Result     []*UserPoolDTO `json:"result,omitempty"`
 }