+++ services/agentidentity/model_update_agent_response.go
@@ -1,0 +1,16 @@
+package agentidentity
+
+import "github.com/baidubce/baiducloud-go-sdk/bce"
+
+type UpdateAgentResponse struct {
+	bce.BaseResponse
+	Id                              *string   `json:"id,omitempty"`
+	BceDomainId                     *string   `json:"bceDomainId,omitempty"`
+	BceUserId                       *string   `json:"bceUserId,omitempty"`
+	Name                            *string   `json:"name,omitempty"`
+	Description                     *string   `json:"description,omitempty"`
+	Extra                           *string   `json:"extra,omitempty"`
+	AllowedResourceOauth2ReturnUrls []*string `json:"allowedResourceOauth2ReturnUrls,omitempty"`
+	CreatedAt                       *string   `json:"createdAt,omitempty"`
+	UpdatedAt                       *string   `json:"updatedAt,omitempty"`
+}